requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==1.10.15
python-multipart==0.0.9
python-dotenv==1.0.0


